Move classes generated from `aggregate.proto` into `aggregate` package
Also:
* Removed unused code from `testdata.Sample`.
