+++ server/src/main/java/org/spine3/server/aggregate/Aggregate.java
@@ -30,8 +30,6 @@ import org.spine3.base.EventContext;
 import org.spine3.base.Version;
 import org.spine3.protobuf.AnyPacker;
 import org.spine3.server.aggregate.error.MissingEventApplierException;
-import org.spine3.server.aggregate.storage.AggregateEvents;
-import org.spine3.server.aggregate.storage.Snapshot;
 import org.spine3.server.command.CommandHandlingEntity;
 import org.spine3.server.entity.Visibility;
 import org.spine3.server.reflect.MethodRegistry;
