+++ server/src/main/proto/spine/server/aggregate/aggregate.proto
@@ -23,10 +23,8 @@ package spine.server.aggregate;
 
 option (type_url_prefix) = "type.spine3.org";
 option (SPI_all) = true;
-option java_package = "org.spine3.server.aggregate.storage";
-    // Put Java classes into sub-package `storage` so that we can annotate
-    // the whole package with @SPI via `package-info.java`.
-option java_outer_classname = "AggregateStorageProto";
+option java_package = "org.spine3.server.aggregate";
+option java_outer_classname = "AggregateProto";
 option java_multiple_files = true;
 option java_generate_equals_and_hash = true;
